backend/app/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Data processing
pandas==2.2.2
openpyxl==3.1.3
xlrd==2.0.1
numpy==1.26.4

# Databases
supabase==2.4.0
psycopg2-binary==2.9.9
# sqlite3 built-in → removed

# Background jobs
rq==1.15.1
redis==5.0.3

# Auth & security
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
httpx==0.25.2
pydantic==2.8.2
pyyaml==6.0.1

# Payments
stripe==12.5.0

# AI / LLM
openai==1.37.0

# Misc
typing-extensions==4.12.2
